refactor: 타입 정리
