+++ apps/audience/src/shared/types/festival.ts
@@ -1,3 +1,22 @@
+export interface Festival {
+  festivalId: number;
+  title: string;
+  mainImageUrl: string;
+  period: string;
+  status: '관람 중' | '관람 예정' | '관람 완료';
+  wishList: boolean;
+  dDay?: number;
+}
+
+export interface PaginationResponse {
+  currentPage: number;
+  totalPages: number;
+  totalElements: number;
+  size: number;
+  hasNext: boolean;
+  hasPrevious: boolean;
+}
+
 export interface WishListRequest {
   wishList: boolean;
 }
