+++ apps/audience/src/shared/types/home-response.ts
@@ -1,51 +1,8 @@
-export interface AllFestivalItem {
-  festivalId: number;
-  title: string;
-  mainImageUrl: string;
-  location: string;
-  period: string;
-  wishList: boolean;
-  dDay: number;
-}
+import { Festival, PaginationResponse } from './festival';
 
-export interface UpcomingFestivalItem {
-  festivalId: number;
-  title: string;
-  mainImageUrl: string;
-  location: string;
-  period: string;
-  status: string;
-  wishList: boolean;
-  dDay: number;
-}
-
-export interface PaginationResponse {
-  currentPage: number;
-  totalPages: number;
-  totalElements: number;
-  size: number;
-  hasNext: boolean;
-  hasPrevious: boolean;
-}
-
-export interface AllFestivalsResponse {
-  festivals: AllFestivalItem[];
+export interface FestivalsResponse {
+  festivals: Festival[];
   pagination: PaginationResponse;
-}
-
-export interface UpcomingFestivalsResponse {
-  festivals: UpcomingFestivalItem[];
-  pagination: PaginationResponse;
-}
-
-export interface UpcomingFestivalResponse {
-  festivalId: number;
-  title: string;
-  mainImageUrl: string;
-  location: string;
-  startDate: string;
-  endDate: string;
-  dday: number;
 }
 
 export interface NicknameResponse {
